Automatic stop

Built-in set of functions to be used in companion with the argument conv_checker in MCMC. These functions are not intended to be used

gelman_convergence(threshold = 1.1, check_invariant = TRUE)

geweke_convergence(threshold = 0.025, check_invariant = TRUE, ...)

auto_convergence()

Arguments

threshold
Numeric value. A Gelman statistic below the threshold will return TRUE.

check_invariant
Logical. When TRUE the function only computes the gelman diagnostic using variables with greater than 1e-10 variance.

...
Further arguments passed to the method.

Value

A function passed to MCMC to check automatic convergence.

Details

In the case of geweke_convergence, threshold sets the p-value for the null \(H0: Z = 0\), i.e. equal means between the first and last chunks of the chain. See coda::geweke.diag. This implies that the higher the threshold, the lower the probability of stopping the chain.

In the case that the chain has more than one parameter, the algorithm will return true if and only if the test fails to reject the null for all the parameters.

The auto_convergence function is the default and is just a wrapper of gelman_convergence and geweke_convergence. This function returns a convergence checker that will be either of the other two depending on wether nchains in MCMC is greater than one--in which case it will use the Gelman test--or not--in which case it will use the Geweke test.